requirements.txt
appnope==0.1.3
asttokens==2.4.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
comm==0.2.0
cryptography==41.0.7
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
executing==2.0.1
flake8==6.1.0
greenlet==3.0.1
idna==3.6
iniconfig==2.0.0
ipykernel==6.27.1
ipython==8.19.0
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.5.1
matplotlib-inline==0.1.6
mccabe==0.7.0
nest-asyncio==1.5.8
packaging==23.2
parso==0.8.3
pexpect==4.9.0
platformdirs==4.1.0
playwright==1.40.0
playwright-stealth==1.0.6
pluggy==1.3.0
prompt-toolkit==3.0.43
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.0
pycparser==2.21
pyee==11.0.1
pyflakes==3.1.0
PyGithub==2.1.1
Pygments==2.17.2
PyJWT==2.8.0
PyNaCl==1.5.0
pytest==7.4.3
pytest-base-url==2.0.0
pytest-playwright==0.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pyzmq==25.1.2
requests==2.31.0
six==1.16.0
stack-data==0.6.3
text-unidecode==1.3
tornado==6.4
traitlets==5.14.0
typing_extensions==4.9.0
urllib3==2.1.0
wcwidth==0.2.12
wrapt==1.16.0
